DEV: Move chat service objects into core
