+++ plugins/chat/app/jobs/service_job.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class ServiceJob < ::Jobs::Base
-  include Chat::WithServiceHelper
+  include WithServiceHelper
 
   def run_service(service, dependencies)
     @_result = service.call(dependencies)
