+++ spec/lib/steps_inspector_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Chat::StepsInspector do
+RSpec.describe StepsInspector do
   class DummyService
     include Service::Base
